Move EmbedVersion to util package
