+++ core/trino-main/src/main/java/io/trino/server/Server.java
@@ -63,7 +63,7 @@ import io.trino.server.security.ServerSecurityModule;
 import io.trino.server.security.oauth2.OAuth2Client;
 import io.trino.spi.connector.CatalogHandle;
 import io.trino.transaction.TransactionManagerModule;
-import io.trino.version.EmbedVersion;
+import io.trino.util.EmbedVersion;
 import org.weakref.jmx.guice.MBeanModule;
 
 import java.io.IOException;
