+++ core/trino-main/src/main/java/io/trino/util/EmbedVersion.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.version;
+package io.trino.util;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
