+++ core/trino-main/src/test/java/io/trino/util/TestEmbedVersion.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.version;
+package io.trino.util;
 
 import org.junit.jupiter.api.Test;
